Added Azure uploading to Azure. Removed node_modules.
