+++ src/other_examples/upload_to_big3_storage/functions/nodejs/host.json
@@ -1,0 +1,7 @@
+{
+  "version": "2.0",
+  "extensionBundle": {
+    "id": "Microsoft.Azure.Functions.ExtensionBundle",
+    "version": "[1.*, 5.0.0)"
+  }
+}
