--- src/other_examples/upload_to_big3_storage/functions/nodejs/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "upload_to_big3_storage",
-  "version": "1.0.0",
-  "description": "Cloud function for uploading a file to the Big 3 cloud storage services (where possible) from each cloud provider using Workload Identity Federation.",
-  "main": "upload_from_aws.js",
-  "scripts": {
-    "lint": "eslint .",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/pumasecurity/nymeria.git"
-  },
-  "keywords": [
-    "aws",
-    "azure",
-    "google",
-    "cloud",
-    "workload",
-    "identity",
-    "federation",
-    "identity",
-    "and",
-    "access",
-    "management",
-    "iam"
-  ],
-  "author": "Brandon Evans (@BrandonMaxEvans) and Eric Johnson (@emjohn20)",
-  "license": "MPL-2.0",
-  "bugs": {
-    "url": "https://github.com/pumasecurity/nymeria/issues"
-  },
-  "homepage": "https://github.com/pumasecurity/nymeria#readme",
-  "dependencies": {
-    "@azure/identity": "3.0.0",
-    "@azure/storage-blob": "12.14.0",
-    "@google-cloud/secret-manager": "5.0.1",
-    "@google-cloud/storage": "7.1.0",
-    "aws-sdk": "2.1459.0",
-    "axios": "1.5.0"
-  }
-}
