+++ src/other_examples/upload_to_big3_storage/functions/nodejs/proxies.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "http://json.schemastore.org/proxies",
+  "proxies": {}
+}
