+++ src/other_examples/upload_to_big3_storage/terraform/aws/function.tf
@@ -1,16 +1,16 @@
 data "archive_file" "upload_to_big3_storage_build" {
   type        = "zip"
-  source_dir  = "${path.module}/../../functions/nodejs"
-  output_path = "${path.module}/../functions/nodejs.zip"
+  source_dir  = "${path.module}/../../functions/nodejs/upload"
+  output_path = "/tmp/upload-to-big3-nodejs-aws.zip"
 }
 
 resource "aws_lambda_function" "upload_to_big3_storage" {
-  function_name    = "upload-to-big3-${var.unique_identifier}"
+  function_name    = "upload-to-big3"
   description      = "Cloud function for uploading a file to the Big 3 cloud storage services (where possible) from each cloud provider using Workload Identity Federation."
   runtime          = "nodejs18.x"
   handler          = "index.handler"
   role             = aws_iam_role.function.arn
-  filename         = "${path.module}/../functions/nodejs.zip"
+  filename         = data.archive_file.upload_to_big3_storage_build.output_path
   source_code_hash = data.archive_file.upload_to_big3_storage_build.output_base64sha256
 
   environment {
