+++ src/other_examples/upload_to_big3_storage/terraform/azure/outputs.tf
@@ -1,0 +1,4 @@
+output "azure_function_host" {
+  description = "Azure function URL endpoint"
+  value       = azurerm_linux_function_app.upload_to_big3_storage.default_hostname
+}
