+++ src/other_examples/upload_to_big3_storage/terraform/azure/providers.tf
@@ -1,0 +1,21 @@
+terraform {
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "3.73.0"
+    }
+
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "2.42.0"
+    }
+  }
+}
+
+provider "azurerm" {
+  features {
+    resource_group {
+      prevent_deletion_if_contains_resources = false
+    }
+  }
+}
