+++ src/other_examples/upload_to_big3_storage/terraform/azure/resourceGroup.tf
@@ -1,0 +1,4 @@
+resource "azurerm_resource_group" "upload_to_big3_storage" {
+  name     = "uploadtobig3"
+  location = var.azure_location
+}
